fix(ui,docs): button
