+++ packages/ui/components/molecules/index.tsx
@@ -1,3 +1,2 @@
 // molecules exports
 export * from "./Button";
-export * from "./NewButton";
